Song: store segments instead of full waveform
It's needed to use a more functional approach to evaluating
a song in a given moment (will come in some next commit).

It's also a step towards giving a possibility to use Song objects
to e.g. graphically render them as musical notes, and generally
treat an instance of Song as multiple smaller parts, not just
one waveform blob.