PlotAssert: implement reading raw visualisation from DSL
PlotAssert is a working name of a small library for asserting
on math functions of one variable. The point is to be able
to check if a given function's shape is more or less the desired
one. For the existing code, it's useful for testing the primitive
wave generators. Later on, it will come in handy for testing
envelopes.

It's the first commit that implements the first step of creting
PlotAssert assertions.

This library can be refactored out and live as a separate project.
For now, for simplicity, it will live here.

A part of #4.