Optimize synthesis time by using bucketing
Roughly speaking, it reduces computational complexity of synthesis
from square to linear. Previously for each sample, the whole list of
track segments were scanned sequentially, until the matching one was
met. Now, the whole song is divided into buckets, 1 second each,
and each bucket contains track segments that are heard in the given
second. Thanks to this, the search is greatly narrowed and for generic
songs, it boils down to checking a few track segments per each sample
instead of donzens or hundreds of them.

Part of #12.