Split synthesizer instrument to allow correct glissando
There is an issue when using glissando with instruments that change
their characteristics over time. An example was the synthesizer with
hardcoded simple envelope. When using it with glissando, its playback
was finished prematurely because of the way glissando is implemented.
Glissando should be implemented so that it doesn't rely on the instrument
at 1 Hz. This change is a mitigation of the issue.

Part of #26.