Fix hot-reloading of web page during development
There's a bug in Kotlin that makes hot-reloading not work properly:
ttps://youtrack.jetbrains.com/issue/KT-32273. The Webpack plugin being added
was found in this bug's comments. It seems to work.

One note: the file has to be actually saved to trigger the reload, so e.g.
window focus has to go away from IntelliJ.

Resolves #56.