Play audio using <audio> tag
It's needed because Wavesurfer.js uses the `<audio>` tag - in fact it's
a nice visual wrapper over this tag. Other benefits of this change are
more control over the playback of the synthesized song, and no need to
resynthesize if one wants to play it again.

There's one problem I noticed: sometimes, the total time shown by the
slider is like 13 hours or so. It's non-deterministic, i.e. subsequent
updates of the audio tag's `src` once give incorrect total time, and once
the correct one.

Part of #55.