Apply tempo adjustment right before synthesis
This refactoring is needed because when Web Worker will be used
for non-blocking synthesis, it's not possible to send the whole
Song object to the worker. I'm planning to send only song's name,
and resolving song's name to the appropriate Song instance will happen
inside the worker. Then also any processing (including tempo
adjustment) has to happen already in the worker. This change moves the
tempo adjustment to the place that will be extracted to the worker.

Part of #14.