Report progress when synthesizing
This is the main change to make the UI more responsive to what's
happening in the background. It's not yet possible to do anything
when the synthesis is in progress (e.g. abort it), but just showing
the progress and not blocking the UI is a big step forward.

Part of #14.